Refactor: Move ChipWooConvertCurrency class out of main plugin file
- Extract ChipWooConvertCurrency class to includes/class-chipwooconvertcurrency.php
- chip-woo-convert-currency.php now serves as a thin bootstrap:
  defines constants, loads the class file, and instantiates the singleton
- Fix __FILE__ reference in before_woocommerce_init action to use CHIP_WCC_FILE
  (so HPOS compatibility declaration still points to the main plugin file)
- Fix require_once paths in set_currency_provider() to be absolute via
  plugin_dir_path( CHIP_WCC_FILE )
- Remove phpcs.xml exclude-pattern for main plugin file (no longer needed)

PHPCS: 0 errors, 6 warnings (unused callback params only).
PHPUnit: 13 tests, 15 assertions, all passing.
